exercises all the way to 4.22 (4.23* missing)

+++ palautusrepositorio/osa4/blogi/controllers/blogs.js
@@ -1,25 +1,58 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+/*
+const getTokenFrom = request => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.startsWith('Bearer ')) {
+        return authorization.replace('Bearer ', '')
+    }
+    return null
+}*/
 
 blogsRouter.get('/', async (request, response) => {
-    const allBlogs = await Blog.find({})
+    const allBlogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+    
     response.json(allBlogs)
   })
   
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = request.user
 
     if (!body.title || !body.url) {
         return response.status(400).end()
     }
-    
-    const blog = new Blog(request.body)
+
+    const blog = new Blog({
+        ...request.body,
+        user: user._id
+    })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)    
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    const user = request.user
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
